feat: implement dual-approach multi-provider and multi-party coordination
This commit implements a dual-approach solution combining ERC-8001
(multi-party coordination) with ERC-8004 (multi-provider management):

ERC-8001 Coordination:
- Multi-party consensus via EIP-712 attestations
- Cryptographic proof before job completion/rejection
- Prevents unilateral decisions

ERC-8004 Multi-Provider:
- Support for multiple providers on same job
- Automatic equal payment distribution
- Provider set validation before funding

Bug Fixes:
- Fixed SEL_FUND selector in BaseACPHook (was missing expectedBudget param)
- Added empty data checks in fund callbacks
- Updated all hook callbacks to use new _preFund/_postFund signatures
New Contracts:
- CombinedMultiProviderCoordinationHook: Combines both features
- MultiProviderHook: Standalone multi-provider management
- ERC8004ProviderRegistry: Reference implementation
- IMultiPartyCoordination: Generic interface (no ERC-8001 dependency)
Tests:
- 49 tests passing across 5 test suites
- Comprehensive integration tests demonstrating both patterns
- Full coverage for multi-provider, coordination, and combined flows

Documentation:
- Updated README.md with multi-provider examples
- Added hook-profiles.md documentation

+++ test/DirectTest.sol
@@ -9,8 +9,9 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract MockERC20 is ERC20 {
     constructor(string memory name, string memory symbol) ERC20(name, symbol) {
-        _mint(msg.sender, 1000000 * 10**18);
+        _mint(msg.sender, 1000000 * 10 ** 18);
     }
+
     function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
@@ -20,16 +21,16 @@ contract DirectTest is Test {
     AgenticCommerceHooked public acp;
     ERC8001 public coordination;
     MockERC20 public token;
-    
+
     address public client;
     address public provider;
     address public evaluator;
     address public arbiter;
     address public treasury;
     uint256 public clientKey;
-    
+
     bytes32 public domainSeparator;
-    
+
     bytes32 public constant AGENT_INTENT_TYPEHASH = keccak256(
         "AgentIntent(bytes32 payloadHash,uint64 expiry,uint64 nonce,address agentId,bytes32 coordinationType,uint256 coordinationValue,address[] participants)"
     );
@@ -39,42 +40,42 @@ contract DirectTest is Test {
         treasury = makeAddr("treasury");
         acp = new AgenticCommerceHooked(address(token), treasury);
         coordination = new ERC8001();
-        
+
         (client, clientKey) = makeAddrAndKey("client");
         (provider,) = makeAddrAndKey("provider");
         (evaluator,) = makeAddrAndKey("evaluator");
         (arbiter,) = makeAddrAndKey("arbiter");
-        
+
         vm.deal(client, 100 ether);
         vm.deal(provider, 100 ether);
-        
-        token.mint(client, 100000 * 10**18);
-        token.mint(provider, 100000 * 10**18);
-        
+
+        token.mint(client, 100000 * 10 ** 18);
+        token.mint(provider, 100000 * 10 ** 18);
+
         vm.prank(client);
         token.approve(address(acp), type(uint256).max);
         vm.prank(provider);
         token.approve(address(acp), type(uint256).max);
-        
+
         domainSeparator = coordination.DOMAIN_SEPARATOR();
     }
 
     function _createParticipants() internal view returns (address[] memory) {
         address[] memory participants = new address[](4);
         address[4] memory addrs = [client, provider, evaluator, arbiter];
-        
-        for (uint i = 0; i < 4; i++) {
-            for (uint j = i + 1; j < 4; j++) {
+
+        for (uint256 i = 0; i < 4; i++) {
+            for (uint256 j = i + 1; j < 4; j++) {
                 if (uint160(addrs[j]) < uint160(addrs[i])) {
                     (addrs[i], addrs[j]) = (addrs[j], addrs[i]);
                 }
             }
         }
-        
-        for (uint i = 0; i < 4; i++) {
+
+        for (uint256 i = 0; i < 4; i++) {
             participants[i] = addrs[i];
         }
-        
+
         return participants;
     }
 
@@ -90,9 +91,9 @@ contract DirectTest is Test {
             metadata: ""
         });
         bytes32 payloadHash = keccak256(abi.encode(payload));
-        
+
         address[] memory participants = _createParticipants();
-        
+
         IERC8001.AgentIntent memory intent = IERC8001.AgentIntent({
             payloadHash: payloadHash,
             expiry: uint64(block.timestamp + 1 hours),
@@ -102,27 +103,29 @@ contract DirectTest is Test {
             coordinationValue: 0,
             participants: participants
         });
-        
+
         // Sign intent
         bytes32 participantsHash = keccak256(abi.encodePacked(intent.participants));
-        bytes32 structHash = keccak256(abi.encode(
-            AGENT_INTENT_TYPEHASH,
-            intent.payloadHash,
-            intent.expiry,
-            intent.nonce,
-            intent.agentId,
-            intent.coordinationType,
-            intent.coordinationValue,
-            participantsHash
-        ));
+        bytes32 structHash = keccak256(
+            abi.encode(
+                AGENT_INTENT_TYPEHASH,
+                intent.payloadHash,
+                intent.expiry,
+                intent.nonce,
+                intent.agentId,
+                intent.coordinationType,
+                intent.coordinationValue,
+                participantsHash
+            )
+        );
         bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(clientKey, digest);
         bytes memory signature = abi.encodePacked(r, s, v);
-        
+
         // Call ERC-8001 directly
         vm.prank(client);
         bytes32 intentHash = coordination.proposeCoordination(intent, signature, payload);
-        
+
         console.log("Success! Intent hash:");
         console.logBytes32(intentHash);
     }
